BerTopic/requirements.txt
# BERTopic议题分析系统依赖包 - 兼容性优化版本
# Python >= 3.9, <= 3.11 (推荐 3.11)
# 所有版本经过兼容性测试，确保互相适配

# ========== 核心ML依赖 (经过兼容性验证) ==========
# BerTopic生态系统
bertopic>=0.15.0,<0.17.0      # 主题建模核心包
sentence-transformers>=2.2.0,<3.0.0  # 文本嵌入模型

# 降维和聚类算法
umap-learn>=0.5.3,<0.6.0      # UMAP降维算法
hdbscan>=0.8.29,<0.9.0        # 分层密度聚类

# ========== 数据处理核心包 ==========
pandas>=1.5.0,<2.1.0          # 数据处理框架
numpy>=1.21.0,<1.25.0         # 数值计算基础
openpyxl>=3.0.0,<3.2.0        # Excel文件读写
pyyaml>=6.0,<7.0               # YAML配置文件

# ========== 机器学习框架 ==========
scikit-learn>=1.2.0,<1.4.0    # 传统机器学习算法

# ========== 深度学习框架 (兼容配置) ==========
# PyTorch生态系统 - 使用兼容版本范围
torch>=1.13.0,<2.2.0          # PyTorch核心
transformers>=4.21.0,<5.0.0   # Hugging Face Transformers

# ========== 可视化包 (稳定版本) ==========
plotly>=5.15.0,<6.0.0         # 交互式图表
matplotlib>=3.6.0,<3.9.0      # 静态图表
seaborn>=0.11.0,<0.14.0       # 统计可视化

# ========== 自然语言处理 ==========
spacy>=3.4.0,<3.8.0           # NLP处理框架
langdetect>=1.0.9,<1.1.0      # 语言检测

# ========== Web界面框架 ==========
streamlit>=1.25.0,<2.0.0      # Web应用框架
streamlit-aggrid>=0.3.4,<1.0.0  # 高级表格组件

# ========== 辅助工具包 ==========
tqdm>=4.64.0,<5.0.0           # 进度条显示
python-dateutil>=2.8.0,<3.0.0  # 日期时间处理

# ========== 科学计算扩展 ==========
scipy>=1.9.0,<1.12.0          # 科学计算库
scikit-image>=0.19.0,<0.22.0  # 图像处理

# ========== 可选加速包 (GPU支持) ==========
# 如果有NVIDIA GPU，可取消注释以下包
# torch-audio>=0.13.0,<2.2.0
# torch-vision>=0.14.0,<0.18.0
# cupy-cuda11x>=11.0.0,<13.0.0  # CUDA加速

# ========== spaCy语言模型 ==========
# python -m spacy download zh_core_web_sm    # 中文模型
# python -m spacy download en_core_web_sm    # 英文模型  
# python -m spacy download ru_core_news_sm   # 俄文模型

# ========== 开发和调试工具 (可选) ==========
# ipython>=8.0.0,<9.0.0       # 交互式Python环境
# jupyter>=1.0.0,<2.0.0       # Jupyter Notebook
# pipdeptree>=2.0.0,<3.0.0    # 依赖树可视化